fix(tests): avoid float to double promotions in two new test files (#28650)
The route projection test assigns the float NAN macro to double latitude
fields and the vision target estimator test assigns a float altitude to
the double altitude_msl_m field. The macOS toolchain rejects both under
-Wdouble-promotion with -Werror, so the test suite stopped building there
after #28248 and #28552. Cast at the three sites, as the other tests do.

+++ src/modules/vision_target_estimator/test/TEST_VTE_VisionTargetEst.cpp
@@ -345,7 +345,7 @@ protected:
 		msg.timestamp_sample = timestamp;
 		msg.latitude_deg = lat;
 		msg.longitude_deg = lon;
-		msg.altitude_msl_m = alt_amsl;
+		msg.altitude_msl_m = static_cast<double>(alt_amsl);
 		msg.fix_type = fix_type;
 		ASSERT_TRUE(_uav_gps_pub->publish(msg));
 	}
